// src/constant/education-type.ts
export interface educationTableType {
  id: string;
  documentId: string;
  insitituion: string;
  course: string;
  description: string;
  date: string;
  location: string;
}

export interface educationType {
  id: number;
  documentId: string;
  insitituion: string;
  course: string;
  description: string;
  date: string;
  location: string;
}